fastmvc_secrets/pyproject.toml
[project]
name = "fastmvc_secrets"
version = "0.1.0"
description = "Secrets backends (Vault, AWS Secrets Manager, GCP) for FastMVC."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmvc_core",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]
vault = ["hvac>=2.0.0"]
aws = ["boto3>=1.28.0"]
gcp = ["google-cloud-secret-manager>=2.0.0"]

[project.urls]
Homepage = "https://github.com/your-org/fastmvc_secrets"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
